Move the rest of the input functions out of main.c
