+++ src/main.c
@@ -33,9 +33,6 @@
 #define ALIGN_CENTER 1
 #define ALIGN_RIGHT 2
 
-#define MODE_OFFLINE 0
-#define MODE_ONLINE 1
-
 #define WORKER_IDLE 0
 #define WORKER_BUSY 1
 #define WORKER_DONE 2
@@ -83,13 +80,6 @@ int get_scale_factor()
     return result;
 }
 
-void get_sight_vector(float rx, float ry, float *vx, float *vy, float *vz)
-{
-    float m = cosf(ry);
-    *vx = cosf(rx - RADIANS(90)) * m;
-    *vy = sinf(ry);
-    *vz = sinf(rx - RADIANS(90)) * m;
-}
 
 GLuint gen_crosshair_buffer()
 {
@@ -435,13 +425,6 @@ Player *player_crosshair(Player *player)
     return result;
 }
 
-int chunk_distance(Chunk *chunk, int p, int q)
-{
-    int dp = ABS(chunk->p - p);
-    int dq = ABS(chunk->q - q);
-    return MAX(dp, dq);
-}
-
 int chunk_visible(float planes[6][4], int p, int q, int miny, int maxy)
 {
     int x = p * CHUNK_SIZE - 1;
@@ -487,152 +470,6 @@ int chunk_visible(float planes[6][4], int p, int q, int miny, int maxy)
         }
     }
     return 1;
-}
-
-int _hit_test(
-    Map *map, float max_distance, int previous,
-    float x, float y, float z,
-    float vx, float vy, float vz,
-    int *hx, int *hy, int *hz)
-{
-    int m = 32;
-    int px = 0;
-    int py = 0;
-    int pz = 0;
-    for (int i = 0; i < max_distance * m; i++)
-    {
-        int nx = roundf(x);
-        int ny = roundf(y);
-        int nz = roundf(z);
-        if (nx != px || ny != py || nz != pz)
-        {
-            int hw = map_get(map, nx, ny, nz);
-            if (hw > 0)
-            {
-                if (previous)
-                {
-                    *hx = px;
-                    *hy = py;
-                    *hz = pz;
-                }
-                else
-                {
-                    *hx = nx;
-                    *hy = ny;
-                    *hz = nz;
-                }
-                return hw;
-            }
-            px = nx;
-            py = ny;
-            pz = nz;
-        }
-        x += vx / m;
-        y += vy / m;
-        z += vz / m;
-    }
-    return 0;
-}
-
-int hit_test(
-    int previous, float x, float y, float z, float rx, float ry,
-    int *bx, int *by, int *bz)
-{
-    int result = 0;
-    float best = 0;
-    int p = chunked(x);
-    int q = chunked(z);
-    float vx, vy, vz;
-    get_sight_vector(rx, ry, &vx, &vy, &vz);
-    for (int i = 0; i < g->chunk_count; i++)
-    {
-        Chunk *chunk = g->chunks + i;
-        if (chunk_distance(chunk, p, q) > 1)
-        {
-            continue;
-        }
-        int hx, hy, hz;
-        int hw = _hit_test(&chunk->map, 8, previous,
-                           x, y, z, vx, vy, vz, &hx, &hy, &hz);
-        if (hw > 0)
-        {
-            float d = sqrtf(
-                powf(hx - x, 2) + powf(hy - y, 2) + powf(hz - z, 2));
-            if (best == 0 || d < best)
-            {
-                best = d;
-                *bx = hx;
-                *by = hy;
-                *bz = hz;
-                result = hw;
-            }
-        }
-    }
-    return result;
-}
-
-int hit_test_face(Player *player, int *x, int *y, int *z, int *face)
-{
-    State *s = &player->state;
-    int w = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, x, y, z);
-    if (is_obstacle(w))
-    {
-        int hx, hy, hz;
-        hit_test(1, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
-        int dx = hx - *x;
-        int dy = hy - *y;
-        int dz = hz - *z;
-        if (dx == -1 && dy == 0 && dz == 0)
-        {
-            *face = 0;
-            return 1;
-        }
-        if (dx == 1 && dy == 0 && dz == 0)
-        {
-            *face = 1;
-            return 1;
-        }
-        if (dx == 0 && dy == 0 && dz == -1)
-        {
-            *face = 2;
-            return 1;
-        }
-        if (dx == 0 && dy == 0 && dz == 1)
-        {
-            *face = 3;
-            return 1;
-        }
-        if (dx == 0 && dy == 1 && dz == 0)
-        {
-            int degrees = roundf(DEGREES(atan2f(s->x - hx, s->z - hz)));
-            if (degrees < 0)
-            {
-                degrees += 360;
-            }
-            int top = ((degrees + 45) / 90) % 4;
-            *face = 4 + top;
-            return 1;
-        }
-    }
-    return 0;
-}
-
-int player_intersects_block(
-    int height,
-    float x, float y, float z,
-    int hx, int hy, int hz)
-{
-    int nx = roundf(x);
-    int ny = roundf(y);
-    int nz = roundf(z);
-    for (int i = 0; i < height; i++)
-    {
-        if (nx == hx && ny - i == hy && nz == hz)
-        {
-            return 1;
-        }
-    }
-    return 0;
 }
 
 int _gen_sign_buffer(
@@ -729,54 +566,6 @@ void gen_sign_buffer(Chunk *chunk)
     del_buffer(chunk->sign_buffer);
     chunk->sign_buffer = gen_faces(5, faces, data);
     chunk->sign_faces = faces;
-}
-
-int has_lights(Chunk *chunk)
-{
-    if (!SHOW_LIGHTS)
-    {
-        return 0;
-    }
-    for (int dp = -1; dp <= 1; dp++)
-    {
-        for (int dq = -1; dq <= 1; dq++)
-        {
-            Chunk *other = chunk;
-            if (dp || dq)
-            {
-                other = find_chunk(chunk->p + dp, chunk->q + dq, g);
-            }
-            if (!other)
-            {
-                continue;
-            }
-            Map *map = &other->lights;
-            if (map->size)
-            {
-                return 1;
-            }
-        }
-    }
-    return 0;
-}
-
-void dirty_chunk(Chunk *chunk)
-{
-    chunk->dirty = 1;
-    if (has_lights(chunk))
-    {
-        for (int dp = -1; dp <= 1; dp++)
-        {
-            for (int dq = -1; dq <= 1; dq++)
-            {
-                Chunk *other = find_chunk(chunk->p + dp, chunk->q + dq, g);
-                if (other)
-                {
-                    other->dirty = 1;
-                }
-            }
-        }
-    }
 }
 
 void occlusion(
@@ -1154,7 +943,7 @@ void init_chunk(Chunk *chunk, int p, int q)
     chunk->sign_faces = 0;
     chunk->buffer = 0;
     chunk->sign_buffer = 0;
-    dirty_chunk(chunk);
+    dirty_chunk(chunk, g);
     SignList *signs = &chunk->signs;
     sign_list_alloc(signs, 16);
     db_load_signs(signs, p, q);
@@ -1454,200 +1243,6 @@ int worker_run(void *arg)
     return 0;
 }
 
-void unset_sign(int x, int y, int z)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        SignList *signs = &chunk->signs;
-        if (sign_list_remove_all(signs, x, y, z))
-        {
-            chunk->dirty = 1;
-            db_delete_signs(x, y, z);
-        }
-    }
-    else
-    {
-        db_delete_signs(x, y, z);
-    }
-}
-
-void unset_sign_face(int x, int y, int z, int face)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        SignList *signs = &chunk->signs;
-        if (sign_list_remove(signs, x, y, z, face))
-        {
-            chunk->dirty = 1;
-            db_delete_sign(x, y, z, face);
-        }
-    }
-    else
-    {
-        db_delete_sign(x, y, z, face);
-    }
-}
-
-void _set_sign(
-    int p, int q, int x, int y, int z, int face, const char *text, int dirty)
-{
-    if (strlen(text) == 0)
-    {
-        unset_sign_face(x, y, z, face);
-        return;
-    }
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        SignList *signs = &chunk->signs;
-        sign_list_add(signs, x, y, z, face, text);
-        if (dirty)
-        {
-            chunk->dirty = 1;
-        }
-    }
-    db_insert_sign(p, q, x, y, z, face, text);
-}
-
-void set_sign(int x, int y, int z, int face, const char *text)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    _set_sign(p, q, x, y, z, face, text, 1);
-    client_sign(x, y, z, face, text);
-}
-
-void toggle_light(int x, int y, int z)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        Map *map = &chunk->lights;
-        int w = map_get(map, x, y, z) ? 0 : 15;
-        map_set(map, x, y, z, w);
-        db_insert_light(p, q, x, y, z, w);
-        client_light(x, y, z, w);
-        dirty_chunk(chunk);
-    }
-}
-
-void set_light(int p, int q, int x, int y, int z, int w)
-{
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        Map *map = &chunk->lights;
-        if (map_set(map, x, y, z, w))
-        {
-            dirty_chunk(chunk);
-            db_insert_light(p, q, x, y, z, w);
-        }
-    }
-    else
-    {
-        db_insert_light(p, q, x, y, z, w);
-    }
-}
-
-void _set_block(int p, int q, int x, int y, int z, int w, int dirty)
-{
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        Map *map = &chunk->map;
-        if (map_set(map, x, y, z, w))
-        {
-            if (dirty)
-            {
-                dirty_chunk(chunk);
-            }
-            db_insert_block(p, q, x, y, z, w);
-        }
-    }
-    else
-    {
-        db_insert_block(p, q, x, y, z, w);
-    }
-    if (w == 0 && chunked(x) == p && chunked(z) == q)
-    {
-        unset_sign(x, y, z);
-        set_light(p, q, x, y, z, 0);
-    }
-}
-
-void set_block(int x, int y, int z, int w)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    _set_block(p, q, x, y, z, w, 1);
-    for (int dx = -1; dx <= 1; dx++)
-    {
-        for (int dz = -1; dz <= 1; dz++)
-        {
-            if (dx == 0 && dz == 0)
-            {
-                continue;
-            }
-            if (dx && chunked(x + dx) == p)
-            {
-                continue;
-            }
-            if (dz && chunked(z + dz) == q)
-            {
-                continue;
-            }
-            _set_block(p + dx, q + dz, x, y, z, -w, 1);
-        }
-    }
-    client_block(x, y, z, w);
-}
-
-void record_block(int x, int y, int z, int w)
-{
-    memcpy(&g->block1, &g->block0, sizeof(Block));
-    g->block0.x = x;
-    g->block0.y = y;
-    g->block0.z = z;
-    g->block0.w = w;
-}
-
-int get_block(int x, int y, int z)
-{
-    int p = chunked(x);
-    int q = chunked(z);
-    Chunk *chunk = find_chunk(p, q, g);
-    if (chunk)
-    {
-        Map *map = &chunk->map;
-        return map_get(map, x, y, z);
-    }
-    return 0;
-}
-
-void builder_block(int x, int y, int z, int w)
-{
-    if (y <= 0 || y >= 256)
-    {
-        return;
-    }
-    if (is_destructable(get_block(x, y, z)))
-    {
-        set_block(x, y, z, 0);
-    }
-    if (w)
-    {
-        set_block(x, y, z, w);
-    }
-}
-
 int render_chunks(Attrib *attrib, Player *player)
 {
     int result = 0;
@@ -1727,7 +1322,7 @@ void render_sign(Attrib *attrib, Player *player)
         return;
     }
     int x, y, z, face;
-    if (!hit_test_face(player, &x, &y, &z, &face))
+    if (!hit_test_face(player, &x, &y, &z, &face, g))
     {
         return;
     }
@@ -1794,7 +1389,7 @@ void render_wireframe(Attrib *attrib, Player *player)
         matrix, g->width, g->height,
         s->x, s->y, s->z, s->rx, s->ry, g->fov, g->ortho, g->render_radius);
     int hx, hy, hz;
-    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
+    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz, g);
     if (is_obstacle(hw))
     {
         glUseProgram(attrib->program);
@@ -1862,711 +1457,6 @@ void render_text(
     del_buffer(buffer);
 }
 
-void add_message(const char *text)
-{
-    printf("%s\n", text);
-    snprintf(
-        g->messages[g->message_index], MAX_TEXT_LENGTH, "%s", text);
-    g->message_index = (g->message_index + 1) % MAX_MESSAGES;
-}
-
-void login()
-{
-    char username[128] = {0};
-    char identity_token[128] = {0};
-    char access_token[128] = {0};
-    if (db_auth_get_selected(username, 128, identity_token, 128))
-    {
-        printf("Contacting login server for username: %s\n", username);
-        if (get_access_token(
-                access_token, 128, username, identity_token))
-        {
-            printf("Successfully authenticated with the login server\n");
-            client_login(username, access_token);
-        }
-        else
-        {
-            printf("Failed to authenticate with the login server\n");
-            client_login("", "");
-        }
-    }
-    else
-    {
-        printf("Logging in anonymously\n");
-        client_login("", "");
-    }
-}
-
-void copy()
-{
-    memcpy(&g->copy0, &g->block0, sizeof(Block));
-    memcpy(&g->copy1, &g->block1, sizeof(Block));
-}
-
-void paste()
-{
-    Block *c1 = &g->copy1;
-    Block *c2 = &g->copy0;
-    Block *p1 = &g->block1;
-    Block *p2 = &g->block0;
-    int scx = SIGN(c2->x - c1->x);
-    int scz = SIGN(c2->z - c1->z);
-    int spx = SIGN(p2->x - p1->x);
-    int spz = SIGN(p2->z - p1->z);
-    int oy = p1->y - c1->y;
-    int dx = ABS(c2->x - c1->x);
-    int dz = ABS(c2->z - c1->z);
-    for (int y = 0; y < 256; y++)
-    {
-        for (int x = 0; x <= dx; x++)
-        {
-            for (int z = 0; z <= dz; z++)
-            {
-                int w = get_block(c1->x + x * scx, y, c1->z + z * scz);
-                builder_block(p1->x + x * spx, y + oy, p1->z + z * spz, w);
-            }
-        }
-    }
-}
-
-void array(Block *b1, Block *b2, int xc, int yc, int zc)
-{
-    if (b1->w != b2->w)
-    {
-        return;
-    }
-    int w = b1->w;
-    int dx = b2->x - b1->x;
-    int dy = b2->y - b1->y;
-    int dz = b2->z - b1->z;
-    xc = dx ? xc : 1;
-    yc = dy ? yc : 1;
-    zc = dz ? zc : 1;
-    for (int i = 0; i < xc; i++)
-    {
-        int x = b1->x + dx * i;
-        for (int j = 0; j < yc; j++)
-        {
-            int y = b1->y + dy * j;
-            for (int k = 0; k < zc; k++)
-            {
-                int z = b1->z + dz * k;
-                builder_block(x, y, z, w);
-            }
-        }
-    }
-}
-
-void cube(Block *b1, Block *b2, int fill)
-{
-    if (b1->w != b2->w)
-    {
-        return;
-    }
-    int w = b1->w;
-    int x1 = MIN(b1->x, b2->x);
-    int y1 = MIN(b1->y, b2->y);
-    int z1 = MIN(b1->z, b2->z);
-    int x2 = MAX(b1->x, b2->x);
-    int y2 = MAX(b1->y, b2->y);
-    int z2 = MAX(b1->z, b2->z);
-    int a = (x1 == x2) + (y1 == y2) + (z1 == z2);
-    for (int x = x1; x <= x2; x++)
-    {
-        for (int y = y1; y <= y2; y++)
-        {
-            for (int z = z1; z <= z2; z++)
-            {
-                if (!fill)
-                {
-                    int n = 0;
-                    n += x == x1 || x == x2;
-                    n += y == y1 || y == y2;
-                    n += z == z1 || z == z2;
-                    if (n <= a)
-                    {
-                        continue;
-                    }
-                }
-                builder_block(x, y, z, w);
-            }
-        }
-    }
-}
-
-void sphere(Block *center, int radius, int fill, int fx, int fy, int fz)
-{
-    static const float offsets[8][3] = {
-        {-0.5, -0.5, -0.5},
-        {-0.5, -0.5, 0.5},
-        {-0.5, 0.5, -0.5},
-        {-0.5, 0.5, 0.5},
-        {0.5, -0.5, -0.5},
-        {0.5, -0.5, 0.5},
-        {0.5, 0.5, -0.5},
-        {0.5, 0.5, 0.5}};
-    int cx = center->x;
-    int cy = center->y;
-    int cz = center->z;
-    int w = center->w;
-    for (int x = cx - radius; x <= cx + radius; x++)
-    {
-        if (fx && x != cx)
-        {
-            continue;
-        }
-        for (int y = cy - radius; y <= cy + radius; y++)
-        {
-            if (fy && y != cy)
-            {
-                continue;
-            }
-            for (int z = cz - radius; z <= cz + radius; z++)
-            {
-                if (fz && z != cz)
-                {
-                    continue;
-                }
-                int inside = 0;
-                int outside = fill;
-                for (int i = 0; i < 8; i++)
-                {
-                    float dx = x + offsets[i][0] - cx;
-                    float dy = y + offsets[i][1] - cy;
-                    float dz = z + offsets[i][2] - cz;
-                    float d = sqrtf(dx * dx + dy * dy + dz * dz);
-                    if (d < radius)
-                    {
-                        inside = 1;
-                    }
-                    else
-                    {
-                        outside = 1;
-                    }
-                }
-                if (inside && outside)
-                {
-                    builder_block(x, y, z, w);
-                }
-            }
-        }
-    }
-}
-
-void cylinder(Block *b1, Block *b2, int radius, int fill)
-{
-    if (b1->w != b2->w)
-    {
-        return;
-    }
-    int w = b1->w;
-    int x1 = MIN(b1->x, b2->x);
-    int y1 = MIN(b1->y, b2->y);
-    int z1 = MIN(b1->z, b2->z);
-    int x2 = MAX(b1->x, b2->x);
-    int y2 = MAX(b1->y, b2->y);
-    int z2 = MAX(b1->z, b2->z);
-    int fx = x1 != x2;
-    int fy = y1 != y2;
-    int fz = z1 != z2;
-    if (fx + fy + fz != 1)
-    {
-        return;
-    }
-    Block block = {x1, y1, z1, w};
-    if (fx)
-    {
-        for (int x = x1; x <= x2; x++)
-        {
-            block.x = x;
-            sphere(&block, radius, fill, 1, 0, 0);
-        }
-    }
-    if (fy)
-    {
-        for (int y = y1; y <= y2; y++)
-        {
-            block.y = y;
-            sphere(&block, radius, fill, 0, 1, 0);
-        }
-    }
-    if (fz)
-    {
-        for (int z = z1; z <= z2; z++)
-        {
-            block.z = z;
-            sphere(&block, radius, fill, 0, 0, 1);
-        }
-    }
-}
-
-void tree(Block *block)
-{
-    int bx = block->x;
-    int by = block->y;
-    int bz = block->z;
-    for (int y = by + 3; y < by + 8; y++)
-    {
-        for (int dx = -3; dx <= 3; dx++)
-        {
-            for (int dz = -3; dz <= 3; dz++)
-            {
-                int dy = y - (by + 4);
-                int d = (dx * dx) + (dy * dy) + (dz * dz);
-                if (d < 11)
-                {
-                    builder_block(bx + dx, y, bz + dz, 15);
-                }
-            }
-        }
-    }
-    for (int y = by; y < by + 7; y++)
-    {
-        builder_block(bx, y, bz, 5);
-    }
-}
-
-void parse_command(const char *buffer, int forward)
-{
-    char username[128] = {0};
-    char token[128] = {0};
-    char server_addr[MAX_ADDR_LENGTH];
-    int server_port = DEFAULT_PORT;
-    char filename[MAX_PATH_LENGTH];
-    int radius, count, xc, yc, zc;
-    if (sscanf(buffer, "/identity %128s %128s", username, token) == 2)
-    {
-        db_auth_set(username, token);
-        add_message("Successfully imported identity token!");
-        login();
-    }
-    else if (strcmp(buffer, "/logout") == 0)
-    {
-        db_auth_select_none();
-        login();
-    }
-    else if (sscanf(buffer, "/login %128s", username) == 1)
-    {
-        if (db_auth_select(username))
-        {
-            login();
-        }
-        else
-        {
-            add_message("Unknown username.");
-        }
-    }
-    else if (sscanf(buffer,
-                    "/online %128s %d", server_addr, &server_port) >= 1)
-    {
-        g->mode_changed = 1;
-        g->mode = MODE_ONLINE;
-        strncpy(g->server_addr, server_addr, MAX_ADDR_LENGTH);
-        g->server_port = server_port;
-        snprintf(g->db_path, MAX_PATH_LENGTH,
-                 "cache.%s.%d.db", g->server_addr, g->server_port);
-    }
-    else if (sscanf(buffer, "/offline %128s", filename) == 1)
-    {
-        g->mode_changed = 1;
-        g->mode = MODE_OFFLINE;
-        snprintf(g->db_path, MAX_PATH_LENGTH, "%s.db", filename);
-    }
-    else if (strcmp(buffer, "/offline") == 0)
-    {
-        g->mode_changed = 1;
-        g->mode = MODE_OFFLINE;
-        snprintf(g->db_path, MAX_PATH_LENGTH, "%s", DB_PATH);
-    }
-    else if (sscanf(buffer, "/view %d", &radius) == 1)
-    {
-        if (radius >= 1 && radius <= 24)
-        {
-            g->create_radius = radius;
-            g->render_radius = radius;
-            g->delete_radius = radius + 4;
-        }
-        else
-        {
-            add_message("Viewing distance must be between 1 and 24.");
-        }
-    }
-    else if (strcmp(buffer, "/copy") == 0)
-    {
-        copy();
-    }
-    else if (strcmp(buffer, "/paste") == 0)
-    {
-        paste();
-    }
-    else if (strcmp(buffer, "/tree") == 0)
-    {
-        tree(&g->block0);
-    }
-    else if (sscanf(buffer, "/array %d %d %d", &xc, &yc, &zc) == 3)
-    {
-        array(&g->block1, &g->block0, xc, yc, zc);
-    }
-    else if (sscanf(buffer, "/array %d", &count) == 1)
-    {
-        array(&g->block1, &g->block0, count, count, count);
-    }
-    else if (strcmp(buffer, "/fcube") == 0)
-    {
-        cube(&g->block0, &g->block1, 1);
-    }
-    else if (strcmp(buffer, "/cube") == 0)
-    {
-        cube(&g->block0, &g->block1, 0);
-    }
-    else if (sscanf(buffer, "/fsphere %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 1, 0, 0, 0);
-    }
-    else if (sscanf(buffer, "/sphere %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 0, 0, 0, 0);
-    }
-    else if (sscanf(buffer, "/fcirclex %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 1, 1, 0, 0);
-    }
-    else if (sscanf(buffer, "/circlex %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 0, 1, 0, 0);
-    }
-    else if (sscanf(buffer, "/fcircley %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 1, 0, 1, 0);
-    }
-    else if (sscanf(buffer, "/circley %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 0, 0, 1, 0);
-    }
-    else if (sscanf(buffer, "/fcirclez %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 1, 0, 0, 1);
-    }
-    else if (sscanf(buffer, "/circlez %d", &radius) == 1)
-    {
-        sphere(&g->block0, radius, 0, 0, 0, 1);
-    }
-    else if (sscanf(buffer, "/fcylinder %d", &radius) == 1)
-    {
-        cylinder(&g->block0, &g->block1, radius, 1);
-    }
-    else if (sscanf(buffer, "/cylinder %d", &radius) == 1)
-    {
-        cylinder(&g->block0, &g->block1, radius, 0);
-    }
-    else if (forward)
-    {
-        client_talk(buffer);
-    }
-}
-
-void on_light()
-{
-    State *s = &g->players->state;
-    int hx, hy, hz;
-    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
-    if (hy > 0 && hy < 256 && is_destructable(hw))
-    {
-        toggle_light(hx, hy, hz);
-    }
-}
-
-void on_left_click()
-{
-    State *s = &g->players->state;
-    int hx, hy, hz;
-    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
-    if (hy > 0 && hy < 256 && is_destructable(hw))
-    {
-        set_block(hx, hy, hz, 0);
-        record_block(hx, hy, hz, 0);
-        if (is_plant(get_block(hx, hy + 1, hz)))
-        {
-            set_block(hx, hy + 1, hz, 0);
-        }
-    }
-}
-
-void on_right_click()
-{
-    State *s = &g->players->state;
-    int hx, hy, hz;
-    int hw = hit_test(1, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
-    if (hy > 0 && hy < 256 && is_obstacle(hw))
-    {
-        if (!player_intersects_block(2, s->x, s->y, s->z, hx, hy, hz))
-        {
-            set_block(hx, hy, hz, items[g->item_index]);
-            record_block(hx, hy, hz, items[g->item_index]);
-        }
-    }
-}
-
-void on_middle_click()
-{
-    State *s = &g->players->state;
-    int hx, hy, hz;
-    int hw = hit_test(0, s->x, s->y, s->z, s->rx, s->ry, &hx, &hy, &hz);
-    for (int i = 0; i < item_count; i++)
-    {
-        if (items[i] == hw)
-        {
-            g->item_index = i;
-            break;
-        }
-    }
-}
-
-void on_key(GLFWwindow *window, int key, int scancode, int action, int mods)
-{
-    int control = mods & (GLFW_MOD_CONTROL | GLFW_MOD_SUPER);
-    int exclusive =
-        glfwGetInputMode(window, GLFW_CURSOR) == GLFW_CURSOR_DISABLED;
-    if (action == GLFW_RELEASE)
-    {
-        return;
-    }
-    if (key == GLFW_KEY_BACKSPACE)
-    {
-        if (g->typing)
-        {
-            int n = strlen(g->typing_buffer);
-            if (n > 0)
-            {
-                g->typing_buffer[n - 1] = '\0';
-            }
-        }
-    }
-    if (action != GLFW_PRESS)
-    {
-        return;
-    }
-    if (key == GLFW_KEY_ESCAPE)
-    {
-        if (g->typing)
-        {
-            g->typing = 0;
-        }
-        else if (exclusive)
-        {
-            glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
-        }
-    }
-    if (key == GLFW_KEY_ENTER)
-    {
-        if (g->typing)
-        {
-            if (mods & GLFW_MOD_SHIFT)
-            {
-                int n = strlen(g->typing_buffer);
-                if (n < MAX_TEXT_LENGTH - 1)
-                {
-                    g->typing_buffer[n] = '\r';
-                    g->typing_buffer[n + 1] = '\0';
-                }
-            }
-            else
-            {
-                g->typing = 0;
-                if (g->typing_buffer[0] == CRAFT_KEY_SIGN)
-                {
-                    Player *player = g->players;
-                    int x, y, z, face;
-                    if (hit_test_face(player, &x, &y, &z, &face))
-                    {
-                        set_sign(x, y, z, face, g->typing_buffer + 1);
-                    }
-                }
-                else if (g->typing_buffer[0] == '/')
-                {
-                    parse_command(g->typing_buffer, 1);
-                }
-                else
-                {
-                    client_talk(g->typing_buffer);
-                }
-            }
-        }
-        else
-        {
-            if (control)
-            {
-                on_right_click();
-            }
-            else
-            {
-                on_left_click();
-            }
-        }
-    }
-    if (control && key == 'V')
-    {
-        const char *buffer = glfwGetClipboardString(window);
-        if (g->typing)
-        {
-            g->suppress_char = 1;
-            strncat(g->typing_buffer, buffer,
-                    MAX_TEXT_LENGTH - strlen(g->typing_buffer) - 1);
-        }
-        else
-        {
-            parse_command(buffer, 0);
-        }
-    }
-    if (!g->typing)
-    {
-        if (key == CRAFT_KEY_FLY)
-        {
-            g->flying = !g->flying;
-        }
-        if (key >= '1' && key <= '9')
-        {
-            g->item_index = key - '1';
-        }
-        if (key == '0')
-        {
-            g->item_index = 9;
-        }
-        if (key == CRAFT_KEY_ITEM_NEXT)
-        {
-            g->item_index = (g->item_index + 1) % item_count;
-        }
-        if (key == CRAFT_KEY_ITEM_PREV)
-        {
-            g->item_index--;
-            if (g->item_index < 0)
-            {
-                g->item_index = item_count - 1;
-            }
-        }
-        if (key == CRAFT_KEY_OBSERVE)
-        {
-            g->observe1 = (g->observe1 + 1) % g->player_count;
-        }
-        if (key == CRAFT_KEY_OBSERVE_INSET)
-        {
-            g->observe2 = (g->observe2 + 1) % g->player_count;
-        }
-    }
-}
-
-void on_char(GLFWwindow *window, unsigned int u)
-{
-    if (g->suppress_char)
-    {
-        g->suppress_char = 0;
-        return;
-    }
-    if (g->typing)
-    {
-        if (u >= 32 && u < 128)
-        {
-            char c = (char)u;
-            int n = strlen(g->typing_buffer);
-            if (n < MAX_TEXT_LENGTH - 1)
-            {
-                g->typing_buffer[n] = c;
-                g->typing_buffer[n + 1] = '\0';
-            }
-        }
-    }
-    else
-    {
-        if (u == CRAFT_KEY_CHAT)
-        {
-            g->typing = 1;
-            g->typing_buffer[0] = '\0';
-        }
-        if (u == CRAFT_KEY_COMMAND)
-        {
-            g->typing = 1;
-            g->typing_buffer[0] = '/';
-            g->typing_buffer[1] = '\0';
-        }
-        if (u == CRAFT_KEY_SIGN)
-        {
-            g->typing = 1;
-            g->typing_buffer[0] = CRAFT_KEY_SIGN;
-            g->typing_buffer[1] = '\0';
-        }
-    }
-}
-
-void on_scroll(GLFWwindow *window, double xdelta, double ydelta)
-{
-    static double ypos = 0;
-    ypos += ydelta;
-    if (ypos < -SCROLL_THRESHOLD)
-    {
-        g->item_index = (g->item_index + 1) % item_count;
-        ypos = 0;
-    }
-    if (ypos > SCROLL_THRESHOLD)
-    {
-        g->item_index--;
-        if (g->item_index < 0)
-        {
-            g->item_index = item_count - 1;
-        }
-        ypos = 0;
-    }
-}
-
-void on_mouse_button(GLFWwindow *window, int button, int action, int mods)
-{
-    int control = mods & (GLFW_MOD_CONTROL | GLFW_MOD_SUPER);
-    int exclusive =
-        glfwGetInputMode(window, GLFW_CURSOR) == GLFW_CURSOR_DISABLED;
-    if (action != GLFW_PRESS)
-    {
-        return;
-    }
-    if (button == GLFW_MOUSE_BUTTON_LEFT)
-    {
-        if (exclusive)
-        {
-            if (control)
-            {
-                on_right_click();
-            }
-            else
-            {
-                on_left_click();
-            }
-        }
-        else
-        {
-            glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
-        }
-    }
-    if (button == GLFW_MOUSE_BUTTON_RIGHT)
-    {
-        if (exclusive)
-        {
-            if (control)
-            {
-                on_light();
-            }
-            else
-            {
-                on_right_click();
-            }
-        }
-    }
-    if (button == GLFW_MOUSE_BUTTON_MIDDLE)
-    {
-        if (exclusive)
-        {
-            on_middle_click();
-        }
-    }
-}
-
 void create_window()
 {
     int window_width = WINDOW_WIDTH;
@@ -2584,45 +1474,56 @@ void create_window()
         window_width, window_height, "Craft", monitor, NULL);
 }
 
-void handle_mouse_input()
+void reset_model()
 {
-    int exclusive =
-        glfwGetInputMode(g->window, GLFW_CURSOR) == GLFW_CURSOR_DISABLED;
-    static double px = 0;
-    static double py = 0;
-    State *s = &g->players->state;
-    if (exclusive && (px || py))
-    {
-        double mx, my;
-        glfwGetCursorPos(g->window, &mx, &my);
-        float m = 0.0025;
-        s->rx += (mx - px) * m;
-        if (INVERT_MOUSE)
-        {
-            s->ry += (my - py) * m;
-        }
-        else
-        {
-            s->ry -= (my - py) * m;
-        }
-        if (s->rx < 0)
-        {
-            s->rx += RADIANS(360);
-        }
-        if (s->rx >= RADIANS(360))
-        {
-            s->rx -= RADIANS(360);
-        }
-        s->ry = MAX(s->ry, -RADIANS(90));
-        s->ry = MIN(s->ry, RADIANS(90));
-        px = mx;
-        py = my;
-    }
-    else
-    {
-        glfwGetCursorPos(g->window, &px, &py);
+    memset(g->chunks, 0, sizeof(Chunk) * MAX_CHUNKS);
+    g->chunk_count = 0;
+    memset(g->players, 0, sizeof(Player) * MAX_PLAYERS);
+    g->player_count = 0;
+    g->observe1 = 0;
+    g->observe2 = 0;
+    g->flying = 0;
+    g->item_index = 0;
+    memset(g->typing_buffer, 0, sizeof(char) * MAX_TEXT_LENGTH);
+    g->typing = 0;
+    memset(g->messages, 0, sizeof(char) * MAX_MESSAGES * MAX_TEXT_LENGTH);
+    g->message_index = 0;
+    g->day_length = DAY_LENGTH;
+    glfwSetTime(g->day_length / 3.0);
+    g->time_changed = 1;
+}
+
+void on_key_wrapper(GLFWwindow *window, int key, int scancode, int action, int mods)
+{
+    if(g){
+        on_key(window, key, scancode, action, mods, g);
     }
 }
+
+void on_char_wrapper(GLFWwindow *window, unsigned int u)
+{
+    if (g)
+    {
+        on_char(window, u, g);
+    }
+}
+
+void on_mouse_button_wrapper(GLFWwindow *window, int button, int action, int mods)
+{
+    if (g)
+    {
+        on_mouse_button(window, button, action, mods, g);
+    }
+}
+
+void on_scroll_wrapper(GLFWwindow *window, double xdelta, double ydelta)
+{
+    if (g)
+    {
+        on_scroll(window, xdelta, ydelta, g);
+    }
+}
+
 void parse_buffer(char *buffer)
 {
     Player *me = g->players;
@@ -2652,7 +1553,7 @@ void parse_buffer(char *buffer)
         if (sscanf(line, "B,%d,%d,%d,%d,%d,%d",
                    &bp, &bq, &bx, &by, &bz, &bw) == 6)
         {
-            _set_block(bp, bq, bx, by, bz, bw, 0);
+            _set_block(bp, bq, bx, by, bz, bw, 0, g);
             if (player_intersects_block(2, s->x, s->y, s->z, bx, by, bz))
             {
                 s->y = highest_block(s->x, s->z, g) + 2;
@@ -2696,7 +1597,7 @@ void parse_buffer(char *buffer)
             Chunk *chunk = find_chunk(kp, kq, g);
             if (chunk)
             {
-                dirty_chunk(chunk);
+                dirty_chunk(chunk, g);
             }
         }
         double elapsed;
@@ -2710,7 +1611,7 @@ void parse_buffer(char *buffer)
         if (line[0] == 'T' && line[1] == ',')
         {
             char *text = line + 2;
-            add_message(text);
+            add_message(text, g);
         }
         char format[64];
         snprintf(
@@ -2732,30 +1633,12 @@ void parse_buffer(char *buffer)
         if (sscanf(line, format,
                    &bp, &bq, &bx, &by, &bz, &face, text) >= 6)
         {
-            _set_sign(bp, bq, bx, by, bz, face, text, 0);
+            _set_sign(bp, bq, bx, by, bz, face, text, 0, g);
         }
         line = tokenize(NULL, "\n", &key);
     }
 }
 
-void reset_model()
-{
-    memset(g->chunks, 0, sizeof(Chunk) * MAX_CHUNKS);
-    g->chunk_count = 0;
-    memset(g->players, 0, sizeof(Player) * MAX_PLAYERS);
-    g->player_count = 0;
-    g->observe1 = 0;
-    g->observe2 = 0;
-    g->flying = 0;
-    g->item_index = 0;
-    memset(g->typing_buffer, 0, sizeof(char) * MAX_TEXT_LENGTH);
-    g->typing = 0;
-    memset(g->messages, 0, sizeof(char) * MAX_MESSAGES * MAX_TEXT_LENGTH);
-    g->message_index = 0;
-    g->day_length = DAY_LENGTH;
-    glfwSetTime(g->day_length / 3.0);
-    g->time_changed = 1;
-}
 
 int main(int argc, char **argv)
 {
@@ -2779,10 +1662,10 @@ int main(int argc, char **argv)
     glfwMakeContextCurrent(g->window);
     glfwSwapInterval(VSYNC);
     glfwSetInputMode(g->window, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
-    glfwSetKeyCallback(g->window, on_key);
-    glfwSetCharCallback(g->window, on_char);
-    glfwSetMouseButtonCallback(g->window, on_mouse_button);
-    glfwSetScrollCallback(g->window, on_scroll);
+    glfwSetKeyCallback(g->window, on_key_wrapper);
+    glfwSetCharCallback(g->window, on_char_wrapper);
+    glfwSetMouseButtonCallback(g->window, on_mouse_button_wrapper);
+    glfwSetScrollCallback(g->window, on_scroll_wrapper);
 
     if (glewInit() != GLEW_OK)
     {
@@ -2933,7 +1816,7 @@ int main(int argc, char **argv)
             client_connect(g->server_addr, g->server_port);
             client_start();
             client_version(1);
-            login();
+            login(g);
         }
 
         // LOCAL VARIABLES //
@@ -2983,7 +1866,7 @@ int main(int argc, char **argv)
             previous = now;
 
             // HANDLE MOUSE INPUT //
-            handle_mouse_input();
+            handle_mouse_input(g);
 
             // HANDLE MOVEMENT //
             handle_movement(dt, g);
